+++ Syadeu/Entity/IStaticDataManager.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Syadeu
+{
+    public interface IStaticDataManager : IStaticManager, IDisposable
+    {
+        bool Disposed { get; }
+    }
+}
