+++ Syadeu/Entity/IStaticDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc95bcc8cb0214c42b7565700c0d2f49
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
