+++ Syadeu/Entity/IStaticMonoManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94bd0166ddfeae64ebce2eab385050d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
