+++ Syadeu/Entity/SystemFlag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77cccb22f1ffe9f438b711fd0d42e661
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
